Bugfixes in cplscheme, reserving space for old data and double conv meas.
Bug 1:
The coupling scheme post-processings did not reserve space for old data
up to now, since always a convergence measure had been defined for the
post-processed data. Since, however, the post-processing is applied to
all send-data, it is not necessary to measure the data given to construct
the post-processing. Now, all post-processings reserve data, if not done
by a convergence measure or extrapolation.

Bug 2:
So far, only one convergence measure for once data could be defined. Now,
arbitrary many can be used.