Removed overflow calculation in precice::cplscheme::AitkenPostProcessing
In the first Aitken iteration, the residual deltas are initialized with
the maximal double value, which are multiplied with the residuals. The
result is not used, however, and the calculation is now not performed
in the first iteration any more. From the second iteration on,
meaningful residual deltas are present already.