Added error message and XML ref. docu for Python action.
When configuring a Python action with wrong target or source data,
now an error message is thrown.

Using library module names for user scripts results in loading the
library module instead of the user script. This behavior is now
documented in the XML reference at the Python module name tag.