Corrected secondary data relaxation in IONILSPostProcessing.
Secondary data is data not involved in computing the IQN coefficients,
but being also relaxed by the IQN coefficients. Before, that data was
wrongly relaxed, using the coupling values of the primary data, which
it not really related if, e.g., the primary data is displacement and
the secondary forces.