Modified/added to Python scripts.
Script insert_mpi_header.py inserts an include of the MPI header into
the first line of every .cpp file, when executed in the root of preCICE.
This can be useful when using the intel MPI compiler wrapper, that has
problems when C standard library headers are included before the MPI
header. The changed files are not meant to be included in the repository.

Script dist.py has been renamed to make_distributable.py and its
documentation been slightly adapted.