Extended SocketCommunication by multi IP and network capabilities.
On a supercomputer, the different compute nodes have different IPs.
Before, socket communication was using the local host address only.
Now, the accepting communication partner writes its IP address into
a file (as for MPI ports communication) and the requesting partner
reads it from there. Thus, communication on a machine with shared
file system but different IPs is now possible.

Second, on machines with different networks it might be necessary
to pich one specific network. This is now possible by utilizing
socket descriptors. The implementation works for BSD sockets, as
found on Linux systems but not on Windows, only.