Some rework of testing infrastructure.
Central problem is that a communicator can only be restricted when all
ranks participate in the restrictions.

Since we filter tests when building the test tree and restrict in a
fixture, this can cause deadlocks.

The way to go, currently, is to use the MinSize or OnSize decorators (PETSc
needs exactly 4 ranks), both which do not delete single ranks (OnSize
either deletes all or none) and to install the fixture manually.

This is work in progress and likely to be improved later.