Building: Exclude some targets from the default list.
Remove the bin and staticlib from the default targets to reduce building time and storage requirements.
In most cases, the shared library should be enough, while the static library takes too much more space.
The binprecice is used mostly for deprecated functionality or for to generate the XML reference,
something that is normally not used by users.