Partial fix for running preCICE without network connection
This partially solves a problem described in #45, where preCICE
fails to run e.g. on a Linux laptop when there is no external network connection.

Without this patch, the problem appears both when running a simulation
and when running the Boost tests with `compileAndTest.py -b`.

With the patch, the problem appears to be solved when running a simulation
(the OpenFOAM-OpenFOAM CHT tutorial, without MPI), but still fails when
running the Boost tests.