Sconstruct and NumPy: Warning instead of error if not imported
The user can still specify custom paths for Python and NumPy.
In this case, not being able to load NumPy by SCons should not be an error,
as we mainly need the header file.
However, this looks like a not so important corner case.
If the header is not found, then an error will be produced by checkAdd().

This commit also adds a few comments and drops the custom short name for
the imported numpy package, as it is only used once.