Integrate the BufferManager class.
Completely decouples the sending from the receiving participant. Instead of sending, immediately followed by a wait call, the request and the buffer is handed off to a separate thread. This thread then tests for the completion of the request and removes the buffer.

Issues remaining.

* Maybe locking in BufferManager::check() is too coarse
* Managed send only integrated in PointToPointCommunication::send()
* Currently only works for std::vector<double>, can be extended to other types by using boost::variant, see code comments and visitor classes.

Occasional crashes are still observed. :-(