Print PRECICE_VERSION as first line of the log.
Output of first three lines:

(0) 23:20:28 []:0 in : This is preCICE version v1.2.0-6-g8f0f2143. Starting 2018-08-08 23-20-28
(0) 23:20:28 []:79 in xmlEndTagCallback: This is preCICE version v1.2.0-6-g8f0f2143. Starting 2018-08-08 23-20-28
(0) 23:20:28 [impl::SolverInterfaceImpl]:115 in configure: Configuring preCICE with configuration: "precice-config.xml"

Yeah, it's in there twice and the meta info is somewhat flaky, but I
think it's fine for the time being.