Call the signal handler on ERROR, put it in it's own cpp file.
Calling the signal handler on error gives us a chance to get the
timings in case of an error.

Put the code in cpp a file, because it is included in LogMacros.hpp,
which in turn is included virtually everywhere. This way, we don't
pull in boost::filesystem und <regexp> everywhere.