Apply const correctness to SolverInterface pointers
Refactor data write functions to take const pointer arguments.
Change the write functions in SolverInterface (and the functions called by them) to take a const double* (resp. const double&) instead of a double* (resp. double) for the value argument. The purpose of these functions is to copy the contents of the buffer provided as value argument to the other solver. Hence, the access to the write buffer is read-only, consequently it can be const.
Add const to the C bindings
Add const in the set/get mesh vertex functions and for the valueIndices in Block functions