Fixing potential memory leak
The pointer to preCICE was never deleted nor was the value of the pointer checked before it was created. I am not sure if it is a valid use case, but a user might finalize preCICE and initialize it again for some reason or somebody calls preCICE's `initialize` functions twice. In such a case I propose that the old object should be either deleted or one should throw error.