Remove mostly redundant _allData. Instead provide it on demand. (#1185)
* Return temporary map with allData, no reference

Assembles a map with all data, when requested. This avoids unnecessarily
storing _allData in the BaseCouplingScheme. Acceleration is still using
references to the data (which does not make really make sense?).

* Remove unnecessary function

BaseCouplingScheme::assignDataToConvergenceMeasure assigns a
PtrCouplingData to convergenceMeasure->couplingData. But this is already
done in BaseCouplingScheme::addConvergenceMeasure at an earlier point in
time. This looks unnecessary.

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>