Increase internal time after setting action.
Previous implementation leads to bug with checkpointing: When entering
the last window, isCouplingOngoing() evaluates false, if we increase the
time before checking isCouplingOngoing(). This will avoid writing a
checkpoint for the beginning of the last window, meaning we jump back to
the second to last window when reading the checkpoing.