Fix bug that prevents requiring checkpoint in last window (#1704)
Previous implementation leads to bug with checkpointing: When entering
the last window, isCouplingOngoing() evaluates false, if we increase the
time before checking isCouplingOngoing() and don't reset the current 
window time to zero. This will avoid writing a checkpoint for the beginning
of the last window, meaning we jump back to the second to last window
when reading the checkpoint.