Release preCICE using annotated tags
Annotated tags include metadata (message, author, timestamp) that tools like [Software Heritage](https://docs.softwareheritage.org/user/faq/index.html#i-cannot-find-all-my-releases-in-a-git-repository-archived-in-software-heritage-what-should-i-do) rely on for properly displaying and recognizing "releases".
Software heritage is an additional archiving platform typically encouraged to use by Darus. preCICE is already there, but the releases are currently not picked-up, so it would be advantageous to have the preCICE releases there by default.