Add function to create Fortran participant using custom communicator
Use MPI_Comm_f2c to convert the Fortran communicator handle to a C handle,
which is then used to create the participant as usual.

Since the precicef_create_with_communicator_ function calls an MPI
library function, if preCICE is built without MPI a compilation error
will result. Added preprocessor directives to fall back to creating
the participant without a communicator in such cases, so that no MPI
is involved. A warning is emitted if the user passes a communicator
when preCICE was built without MPI.

See issue #479