Add HiRSE Code Promotion Badge to README
preCICE is already listed in https://www.helmholtz-hirse.de/promo.html, next to 100+ other projects developed in Germany. The people from HiRSE recently sent around a badge that we could add in our `README.md`.

This is not any kind of achievement, it just highlights that preCICE is developed (at least largely) in Germany, and attracts attention to this list of packages and the packages themselves.

I am just opening this to communicate that it exists, and the badge is available. I think it at least does not harm to have it.